% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonwriter.R
\name{write_json}
\alias{write_json}
\title{Write regression results to a JSON file.}
\usage{
write_json(model, output_filename, verbose = FALSE, extra_fields = NULL, ...)
}
\arguments{
\item{model}{Object that contains the regression results.  Currently
only 'fixest' is supported.}

\item{output_filename}{Filename to write the JSON output to.}

\item{verbose}{Flag to indicate whether to write console messages.
(Optional.)}

\item{extra_fields}{A 'list' object with extra information to add to
the output.  (Optional.)}

\item{...}{Additional arguments passed on to the method.}
}
\value{
Nothing.
}
\description{
Write regression results to a JSON file.
}
\examples{
\dontrun{library(fixest)
data(quakes)

write_json(
    feols(depth ~ mag, quakes),
    "jsonwriter_example.json",
    extra_fields = list(hello = "world"))
}
}
