% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonwriter.R
\name{write_json}
\alias{write_json}
\title{Write regression results to a JSON file.}
\usage{
write_json(model, output_filename, ...)
}
\arguments{
\item{model}{Object that contains the regression results.  Currently
only \code{fixest} is supported.}

\item{output_filename}{Filename to write the JSON output to.}

\item{...}{Additional arguments passed on to the method.}
}
\value{
Nothing.
}
\description{
Write regression results to a JSON file.
}
